Moved remove_with_caveats  to make_sample.py
