+++ runner/operator/roslin_operator/bin/pair_request.py
@@ -5,28 +5,6 @@ import logging
 logger = logging.getLogger(__name__)
 
 
-def remove_with_caveats(samples):
-    data = list()
-    error_data = list()
-    for sample in samples:
-        add = True
-        igo_id = sample['igo_id']
-        sample_name = sample['SM']
-        patient_id = sample['patient_id']
-        if sample_name == 'sampleNameMalformed':
-            add = False
-            logging.debug("Sample name is malformed for for %s; removing from set" % igo_id)
-        if patient_id[:2].lower() not in 'c-':
-            add = False
-            logging.debug("Patient ID does not start with expected 'C-' prefix for %s; removing from set" % igo_id)
-        if add:
-            data.append(sample)
-        else:
-            error_data.append(sample)
-
-    return data, error_data
-
-
 def get_by_tumor_type(data, tumor_type):
     samples = list()
     for sample in data:
